test(canary-web-components): added e2e data table tests
Added more e2e data tables tests which include truncation
